JavaDoc's were updated, small refactoring
