+++ README.md
@@ -18,7 +18,9 @@ There are already [multiple clients for the Nakadi REST API](https://zalando.git
 
 We solved them by persisting new events in a log table as part of the producing JDBC transaction. They will then be sent asynchronously to Nakadi after the transaction completed. If the transaction is rolled back, the events will vanish too. As a result, events will always be sent if and only if the transaction succeeded.
 
-The Transmitter generates a strictly monotonically increasing event id that can be used for ordering the events during retrieval. It is not guaranteed, that events will be sent to Nakadi in the order they have been produced. If an event could not be sent to Nakadi, the library will periodically retry the transmission.
+Unless a different [EID generation strategy](#eid-generation-strategy-optional) is used, the Transmitter generates a strictly monotonically increasing event id that can be used for ordering the events during retrieval.
+
+It is not guaranteed, that events will be sent to Nakadi in the order they have been produced. If an event could not be sent to Nakadi, the library will periodically retry the transmission.
 
 This project is mature, used in production in some services at Zalando, and in active development.
 
@@ -314,7 +316,8 @@ public SnapshotEventGenerator snapshotEventGenerator(MyService service) {
 ```
 
 ### EID generation strategy (optional)
-EID - is unique identifier for nakadi events. We must provide it for each event. By default, the library generates EID based on id of the event in the database.
+The `eid` is a unique identifier for Nakadi events, which is required by Nakadi for each event on submission.
+By default, the library generates eid based on a database sequence (which is also used internally to identify the event log entries before sending them out).
 
 You can implement your own strategy for EID generation by implementing [`EidGeneratorStrategy`](nakadi-producer/src/main/java/org/zalando/nakadiproducer/eventlog/EidGeneratorStrategy.java) interface and providing it as a bean in your application context.
 The library provides some implementations of this interface out of the box:
